style: select animal 데이터 호출 위치 변경
reply page에서 미리 데이터 가져오도록 변경
